Yassmin111111111111111

+++ neurithm/lib/screens/viewAppRatings.dart
@@ -37,7 +37,7 @@ Future<void> _fetchAppRatings() async {
   for (var doc in ratingsQuery.docs) {
     var ratingData = doc.data();
     String patientId = ratingData['patientId'];
-    double rating = ratingData['rating'];
+    int rating = ratingData['rating'];
     var submittedAt = ratingData['submittedAt'];
 
     DateTime submittedDate;
